Add multiple fixes
1. Changes to `get_one_hot`
Problems are given in:
- https://github.com/bioinf-jku/FCD/issues/14
- https://github.com/bioinf-jku/FCD/pull/17
- https://github.com/bioinf-jku/FCD/pull/13

I discarded the changes in the PRs and and added more comprehensive handling of the input data in the
`SmilesDataset` class and the `get_one_hot` function.

2. Imaginary components
Frechet distance calculation fails to work for some cases because of badly conditioned matrices,
as described here https://github.com/bioinf-jku/FCD/issues/15.

Could not reproduce the error locally, but could do so on colab.

Fixed it in `calculate_frechet_distance` by checking if the first `covmean` computation  is real add a small value to the diagonal.
This made it work for me and I got the same result as the original implementation run locally.

3. Added some more tests and changed to pytest

4. As described in https://github.com/bioinf-jku/FCD/pull/16
I changed the data type of the activations to float32 in the `get_predictions` function,
which saves memory for larger datasets.

5. Change to pyproject.toml